created map class, moved map related code to Map.swift
